Robot Final
imported set into robot animation file, fiddled so the walk went to the end of the hall, and bayum. It looks kinda ok in sketchfab, so I don't know if I want to remake it all like I wanted to earlier hrmm, i still may attempt, but I'm pretty ok with it; gave it the good ole college try. First attempt with a lot of things in 3D.